requirements.txt
absl-py==2.1.0albucore==0.0.21albumentations==1.4.23altair==5.5.0annotated-types==0.7.0anyio==4.7.0astunparse==1.6.3attrs==24.3.0beautifulsoup4==4.12.3blinker==1.9.0cachetools==5.5.0certifi==2024.12.14charset-normalizer==3.4.0click==8.1.7contourpy==1.3.1cycler==0.12.1deepface==0.0.93dlib==19.24.6eval_type_backport==0.2.0exceptiongroup==1.2.2fastapi==0.115.6filelock==3.16.1fire==0.7.0Flask==3.1.0Flask-Cors==5.0.0flatbuffers==24.3.25fonttools==4.55.3fsspec==2024.12.0gast==0.6.0gdown==5.2.0gitdb==4.0.11GitPython==3.1.43google-pasta==0.2.0greenlet==3.1.1grpcio==1.68.1gunicorn==23.0.0h11==0.14.0h5py==3.12.1idna==3.10imutils==0.5.4itsdangerous==2.2.0Jinja2==3.1.4joblib==1.4.2jsonschema==4.23.0jsonschema-specifications==2024.10.1keras==3.7.0kiwisolver==1.4.7lap==0.5.12libclang==18.1.1lz4==4.3.3Markdown==3.7markdown-it-py==3.0.0MarkupSafe==3.0.2matplotlib==3.10.0mdurl==0.1.2ml-dtypes==0.4.1mpmath==1.3.0mtcnn==1.0.0namex==0.0.8narwhals==1.19.0networkx==3.4.2numpy==2.0.2nvidia-cublas-cu12==12.5.3.2nvidia-cuda-cupti-cu12==12.5.82nvidia-cuda-nvcc-cu12==12.5.82nvidia-cuda-nvrtc-cu12==12.5.82nvidia-cuda-runtime-cu12==12.5.82nvidia-cudnn-cu12==9.3.0.75nvidia-cufft-cu12==11.2.3.61nvidia-curand-cu12==10.3.6.82nvidia-cusolver-cu12==11.6.3.83nvidia-cusparse-cu12==12.5.1.3#nvidia-nccl-cu12==2.21.5nvidia-nvjitlink-cu12==12.5.82nvidia-nvtx-cu12==12.4.127opencv-contrib-python==4.10.0.84opencv-python==4.10.0.84opencv-python-headless==4.10.0.84opt_einsum==3.4.0optree==0.13.1packaging==24.2pandas==2.2.3pillow==11.0.0pip==22.0.2protobuf==5.29.2psutil==6.1.1psycopg2==2.9.10py-cpuinfo==9.0.0pyarrow==18.1.0pydantic==2.10.4pydantic_core==2.27.2pydeck==0.9.1Pygments==2.18.0pyparsing==3.2.0PySocks==1.7.1python-dateutil==2.9.0.post0pytz==2024.2PyYAML==6.0.2referencing==0.35.1requests==2.32.3retina-face==0.0.17rich==13.9.4rpds-py==0.22.3scikit-learn==1.6.0scipy==1.14.1seaborn==0.13.2setuptools==59.6.0simsimd==6.2.1six==1.17.0smmap==5.0.1sniffio==1.3.1soupsieve==2.6SQLAlchemy==2.0.36starlette==0.41.3streamlit==1.41.1stringzilla==3.11.2sympy==1.13.1tenacity==9.0.0tensorboard==2.18.0tensorboard-data-server==0.7.2tensorflow==2.18.0#tensorflow-io-gcs-filesystem==0.37.1termcolor==2.5.0tf_keras==2.18.0threadpoolctl==3.5.0toml==0.10.2torch==2.5.1torchvision==0.20.1tornado==6.4.2tqdm==4.67.1#triton==3.1.0typing_extensions==4.12.2tzdata==2024.2ultralytics==8.3.52ultralytics-thop==2.0.13urllib3==2.2.3uvicorn==0.34.0watchdog==6.0.0Werkzeug==3.1.3wheel==0.45.1wrapt==1.17.0